# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.AcceptedResponse:
    properties:
      correlation_id:
        type: string
      event:
        type: string
      status:
        type: string
    type: object
...
  handlers.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  handlers.LoginResponse:
    properties:
      api_key:
        type: string
      expires_in_seconds:
        type: integer
    type: object
  handlers.LogoutResponse:
    properties:
      status:
        type: string
    type: object
...
  handlers.UpdateAdminRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    type: object
...
info:
  contact: {}
  description: Event-driven Go API backed by MongoDB and Redis. Every request carries
    a correlation ID (X-Correlation-ID), generated by the server if the caller doesn't
    supply one.
  title: Metarr API
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Compares the submitted username/password against the stored admin
        credentials and issues a session API key carrying admin rights.
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LoginResponse'
        "400":
          description: invalid request body
          schema:
            type: string
        "401":
          description: invalid username or password
          schema:
            type: string
        "429":
          description: too many requests
          schema:
            type: string
      summary: Log in as the admin user
      tags:
      - Auth
  /api/auth/logout:
    post:
      description: Revokes the session API key that authenticated this request.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.LogoutResponse'
        "429":
          description: too many requests
          schema:
            type: string
      security:
      - ApiKeyHeaderAuth: []
      - ApiKeyQueryAuth: []
      summary: Log out
      tags:
      - Auth
...
  /api/config/admin:
    put:
      consumes:
      - application/json
      description: Updates any subset of the admin user's username, email, and password.
        A provided field cannot be empty. If password is set, it is re-hashed with
        a fresh salt.
      parameters:
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateAdminRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/handlers.AcceptedResponse'
        "400":
          description: invalid request body, or a provided field was empty
          schema:
            type: string
      security:
      - ApiKeyHeaderAuth: []
      - ApiKeyQueryAuth: []
      summary: Update the admin user's credentials
      tags:
      - Config
...
securityDefinitions:
  ApiKeyHeaderAuth:
    description: API key via the X-Api-Key header.
    in: header
    name: X-Api-Key
    type: apiKey
  ApiKeyQueryAuth:
    description: API key via the apikey query parameter. Either this or ApiKeyHeaderAuth
      satisfies auth on protected routes.
    in: query
    name: apikey
    type: apiKey
swagger: "2.0"